Remove Imagen samples and add Nano banana image generation (#165)
