+++ app/build.gradle.kts
@@ -86,8 +86,7 @@ dependencies {
     implementation(project(":samples:genai-summarization"))
     implementation(project(":samples:genai-image-description"))
     implementation(project(":samples:genai-writing-assistance"))
-    implementation(project(":samples:imagen"))
-    implementation(project(":samples:imagen-editing"))
+    implementation(project(":samples:nanobanana"))
     implementation(project(":samples:magic-selfie"))
     implementation(project(":samples:gemini-video-summarization"))
     implementation(project(":samples:gemini-live-todo"))
