+++ app/src/main/java/com/android/ai/catalog/domain/SampleCatalog.kt
@@ -32,8 +32,7 @@ import com.android.ai.samples.genai_image_description.GenAIImageDescriptionScree
 import com.android.ai.samples.genai_summarization.GenAISummarizationScreen
 import com.android.ai.samples.genai_writing_assistance.GenAIWritingAssistanceScreen
 import com.android.ai.samples.geminihybrid.GeminiHybridScreen
-import com.android.ai.samples.imagen.ui.ImagenScreen
-import com.android.ai.samples.imagenediting.ui.ImagenEditingScreen
+import com.android.ai.samples.nanobanana.ui.NanobananaScreen
 import com.android.ai.samples.magicselfie.ui.MagicSelfieScreen
 import com.android.ai.theme.extendedColorScheme
 import com.google.firebase.ai.type.PublicPreviewAPI
@@ -60,15 +59,6 @@ val sampleCatalog = listOf(
         keyArt = R.drawable.img_keyart_chatbot,
         needsFirebase = true,
         isFeatured = true,
-    ),
-    SampleCatalogItem(
-        title = R.string.imagen_editing_sample_list_title,
-        description = R.string.imagen_editing_sample_list_description,
-        route = "ImagenMaskEditing",
-        sampleEntryScreen = { ImagenEditingScreen() },
-        tags = listOf(SampleTags.IMAGEN, SampleTags.FIREBASE),
-        needsFirebase = true,
-        keyArt = R.drawable.img_keyart_imagen,
     ),
     SampleCatalogItem(
         title = R.string.gemini_multimodal_sample_list_title,
@@ -114,11 +104,11 @@ val sampleCatalog = listOf(
         keyArt = R.drawable.img_keyart_text,
     ),
     SampleCatalogItem(
-        title = R.string.imagen_sample_list_title,
-        description = R.string.imagen_sample_list_description,
-        route = "ImagenImageGenerationScreen",
-        sampleEntryScreen = { ImagenScreen() },
-        tags = listOf(SampleTags.IMAGEN, SampleTags.FIREBASE),
+        title = R.string.nanobanana_sample_list_title,
+        description = R.string.nanobanana_sample_list_description,
+        route = "NanobananaImageGenerationScreen",
+        sampleEntryScreen = { NanobananaScreen() },
+        tags = listOf(SampleTags.GEMINI_FLASH, SampleTags.FIREBASE),
         needsFirebase = true,
         keyArt = R.drawable.img_keyart_imagen,
     ),
