+++ app/src/main/res/values/strings.xml
@@ -13,10 +13,8 @@
     <string name="top_bar_title">Android AI Samples</string>
     <string name="top_bar_title_expanded">Android\nAI Samples</string>
     <string name="open_sample_button">Open sample</string>
-    <string name="imagen_sample_list_title">Image generation with Imagen</string>
-    <string name="imagen_sample_list_description">Generate images with Imagen, Google image generation model</string>
-    <string name="imagen_editing_sample_list_title">Image Editing with Imagen</string>
-    <string name="imagen_editing_sample_list_description">Generate images and edit only specific areas of a generated image with inpainting</string>
+    <string name="nanobanana_sample_list_title">Image generation with Nanobanana</string>
+    <string name="nanobanana_sample_list_description">Generate images with Nanobanana, Google image generation model</string>
     <string name="magic_selfie_sample_list_title">Magic Selfie with Gemini</string>
     <string name="magic_selfie_sample_list_description">Change the background of your selfies with the Gemini Flash model</string>
     <string name="gemini_video_summarization_sample_list_title">Video Summarization with Gemini and Firebase</string>
