--- samples/imagen/src/main/java/com/android/ai/samples/imagen/ui/GeneratedContent.kt
@@ -1,66 +1,0 @@
-/*
- * Copyright 2025 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.android.ai.samples.imagen.ui
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.material3.Card
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import com.android.ai.samples.imagen.R
-
-@Composable
-fun GeneratedContent(uiState: ImagenUIState, modifier: Modifier = Modifier) {
-    Card(
-        modifier = modifier,
-    ) {
-        when (uiState) {
-            ImagenUIState.Initial -> {
-                //
-            }
-
-            ImagenUIState.Loading -> {
-                //
-            }
-
-            is ImagenUIState.ImageGenerated -> {
-                Image(
-                    bitmap = uiState.bitmap.asImageBitmap(),
-                    contentDescription = uiState.contentDescription,
-                    contentScale = ContentScale.Fit,
-                    modifier = Modifier.fillMaxSize(),
-                )
-            }
-
-            is ImagenUIState.Error -> {
-                Text(
-                    text = uiState.message ?: stringResource(R.string.error_message_unknown),
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .wrapContentSize(Alignment.Center),
-                    textAlign = TextAlign.Center,
-                )
-            }
-        }
-    }
-}
