--- samples/imagen/src/main/res/values/strings.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright 2025 The Android Open Source Project
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~      https://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  ~
-  -->
-
-<resources>
-    <string name="see_code">See Code</string>
-    <string name="placeholder_prompt">An oil painting of Alcatraz</string>
-    <string name="title_image_generation_screen">Imagen image generation</string>
-    <string name="subtitle_image_generation_screen">Generate images with Imagen, Google image generation model</string>
-    <string name="generate_button">Generate</string>
-    <string name="generating_label">Generating…</string>
-    <string name="prompt_label">Prompt</string>
-    <string name="imagen_placeholder">Enter a prompt and tap \"Generate\" to generate an image</string>
-    <string name="error_message_unknown">Unknown error</string>
-</resources>
